remove enable-jobs to create docker image
