+++ conf/application.conf
@@ -151,8 +151,8 @@ features {
   taskReopenAllowedInSeconds = 30
   allowDeleteDatasets = true
   # to enable jobs for local development, use "yarn enable-jobs" to also activate it in the database
-  jobsEnabled = true
-  voxelyticsEnabled = true
+  jobsEnabled = false
+  voxelyticsEnabled = false
   # For new users, the dashboard will show a banner which encourages the user to check out the following dataset.
   # If isWkorgInstance == true, `/createExplorative/hybrid/true` is appended to the URL so that a new tracing is opened.
   # If isWkorgInstance == false, `/view` is appended to the URL so that it's opened in view mode (since the user might not
